% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drcanalyser-package.R
\docType{package}
\name{drcanalyser-package}
\alias{drcanalyser}
\alias{drcanalyser-package}
\title{drcanalyser: Analyse Cell-Viability / Dose-Response Assays}
\description{
Reads Tecan SparkControl plate-reader exports, normalises intensities (IC/GI/GR), fits drc dose-response curves, and produces ggplot figures plus IC50/GI50/GR50 statistics.
}
\author{
\strong{Maintainer}: Ben King \email{bg.king@hotmail.com}

Authors:
\itemize{
  \item Ben King \email{bg.king@hotmail.com}
}

}
\keyword{internal}
