% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{exclude_concentrations}
\alias{exclude_concentrations}
\title{Exclude concentrations from a layout}
\usage{
exclude_concentrations(layout, exclude)
}
\arguments{
\item{layout}{Layout data.frame.}

\item{exclude}{List of \code{c(row, col)} vectors identifying wells to exclude.}
}
\value{
The layout with the specified wells marked.
}
\description{
Marks wells to be removed from a layout by appending \code{"exclude"} to their
label.
}
\keyword{internal}
