% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{format_standard_xl}
\alias{format_standard_xl}
\title{Read a fixed-offset SparkControl plate export}
\usage{
format_standard_xl(path)
}
\arguments{
\item{path}{Path to the Excel file.}
}
\value{
A data.frame of the plate with row labels A:H.
}
\description{
Takes a standard format Excel output from SparkControl and returns a
data.frame of the plate. The plate format assumes 96-wells, has columns 1:12 and rownames A:H.
}
