% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{get_halflog_doses}
\alias{get_halflog_doses}
\title{Compute half-log serial dilution doses}
\usage{
get_halflog_doses(top_dose = 30, n_steps = 0:11)
}
\arguments{
\item{top_dose}{Numeric, the highest dose. Defaults to 30.}

\item{n_steps}{Integer vector of dilution steps. Defaults to \code{0:11}.}
}
\value{
Numeric vector of doses, one per step.
}
\description{
Calculates serial dilution doses for a half-log dilution, with \code{n_steps} and
a \code{top_dose}.
}
