% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{gi}
\alias{gi}
\title{Normalise intensities for a GI50 curve}
\usage{
gi(intensity, negative_control, zero_control)
}
\arguments{
\item{intensity}{Numeric vector of intensities representing drugged wells.}

\item{negative_control}{Average of DMSO-treated (NC-1) wells.}

\item{zero_control}{Average of untreated (NC-0) wells from the GI50 plate.}
}
\value{
Numeric vector of normalised intensities.
}
\description{
\code{y = (y_i - y_0) / (y_nc - y_0) * 100}
}
