% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{gr}
\alias{gr}
\title{Normalise intensities to growth rate}
\usage{
gr(intensity, negative_control, zero_control)
}
\arguments{
\item{intensity}{Numeric vector of intensities representing drugged wells.}

\item{negative_control}{Average of DMSO-treated (NC-1) wells.}

\item{zero_control}{Average of untreated (NC-0) wells from the GI50 plate.}
}
\value{
Numeric vector of growth-rate normalised values.
}
\description{
Normalises data to growth rate, as in Hafner et al., 2016:
\code{y = 2^(log2(y_i/y_0) / log2(y_nc/y_0)) - 1}.
}
