% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{ic}
\alias{ic}
\title{Normalise intensities for an IC50 curve}
\usage{
ic(intensity, negative_control)
}
\arguments{
\item{intensity}{Numeric vector of intensities representing drugged wells.}

\item{negative_control}{Average of DMSO-treated (NC-1) wells.}
}
\value{
Numeric vector of normalised intensities.
}
\description{
\code{y = (y_i / y_nc) * 100}
}
