% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{process_model}
\alias{process_model}
\title{Summarise a fitted model}
\usage{
process_model(model)
}
\arguments{
\item{model}{A fitted \code{drc::drm} model.}
}
\value{
A list with the \code{model}, its \code{summary}, \code{aic}, \code{resid}, \code{ssr} and
\code{mse}.
}
\description{
Processes a model to get statistics (AIC, MSE, etc.).
}
