% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_plates}
\alias{process_plates}
\title{Process a list of plates and fit a model}
\usage{
process_plates(
  path_list,
  assay_id = NULL,
  normalisation_method = c("IC", "GI", "GR")
)
}
\arguments{
\item{path_list}{List of lists containing path information for the plates. In
its most basic form, \code{list(list(main_path = ''), list(main_path = ''))}.}

\item{assay_id}{Optional, appends an \code{assay_id} to the returned processed
plate.}

\item{normalisation_method}{One of \code{"IC"}, \code{"GI"}, \code{"GR"}; selects which
normalisation column the model is fit on.}
}
\value{
A list containing the combined \code{plate} data.frame and the fitted
\code{model}.
}
\description{
Wrapper for \code{\link[=process_plate]{process_plate()}}, for processing lists of plates and fitting a
single dose-response model.
}
