% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{save_results}
\alias{save_results}
\title{Save a plot and statistics to a folder}
\usage{
save_results(
  plot_results,
  save_folder,
  append_file_name = NULL,
  width = 5,
  height = 5
)
}
\arguments{
\item{plot_results}{Results from \code{\link[=plot_drc]{plot_drc()}} or a list containing \verb{$plot}
and \verb{$stats} slots.}

\item{save_folder}{Folder to save results into.}

\item{append_file_name}{String to append to the end of the filename before
the extension.}

\item{width}{Plot width in inches.}

\item{height}{Plot height in inches.}
}
\value{
Invisibly \code{NULL}; writes an pdf plot and a CSV of stats to disk.
}
\description{
Save a plot and statistics to a folder
}
